+++ packages/web/src/views/swap/Swap.tsx
@@ -2,8 +2,9 @@ import React from "react";
 
 import Footer from "@components/common/footer/Footer";
 import HeaderContainer from "@containers/header-container/HeaderContainer";
-import SwapContainer from "@containers/swap-container/SwapContainer";
-import SwapLiquidityContainer from "@containers/swap-liquidity-container/SwapLiquidityContainer";
+
+import SwapContainer from "./containers/swap-container/SwapContainer";
+import SwapLiquidityContainer from "./containers/swap-liquidity-container/SwapLiquidityContainer";
 
 import SwapLayout from "./SwapLayout";
