+++ packages/web/src/views/swap/components/swap-card/swap-card-content/SwapCardContent.tsx
@@ -1,23 +1,25 @@
+import BigNumber from "bignumber.js";
 import React, { useCallback, useMemo } from "react";
+import { useTranslation } from "react-i18next";
+
+import { isAmount } from "@common/utils/data-check-util";
+import IconSwapArrowDown from "@components/common/icons/IconSwapArrowDown";
+import { IconTriangleWarningOutlined } from "@components/common/icons/IconTriangleWarningOutlined";
+import SelectPairButton from "@components/common/select-pair-button/SelectPairButton";
+import SwapCardContentDetail from "@components/swap/swap-card-content-detail/SwapCardContentDetail";
+import { useTheme } from "@emotion/react";
+import { PriceImpactStatus } from "@hooks/swap/use-swap-handler";
+import { SwapRouteInfo } from "@models/swap/swap-route-info";
+import { SwapSummaryInfo } from "@models/swap/swap-summary-info";
+import { SwapTokenInfo } from "@models/swap/swap-token-info";
+import { TokenModel } from "@models/token/token-model";
+
 import {
   ContentWrapper,
   PriceImpactWrapper,
   PriceInfoWrapper,
-  SwapDetailSectionWrapper,
+  SwapDetailSectionWrapper
 } from "./SwapCardContent.styles";
-import IconSwapArrowDown from "@components/common/icons/IconSwapArrowDown";
-import SwapCardContentDetail from "../swap-card-content-detail/SwapCardContentDetail";
-import { SwapTokenInfo } from "@models/swap/swap-token-info";
-import { SwapSummaryInfo } from "@models/swap/swap-summary-info";
-import { SwapRouteInfo } from "@models/swap/swap-route-info";
-import { TokenModel } from "@models/token/token-model";
-import { isAmount } from "@common/utils/data-check-util";
-import SelectPairButton from "@components/common/select-pair-button/SelectPairButton";
-import BigNumber from "bignumber.js";
-import { IconTriangleWarningOutlined } from "@components/common/icons/IconTriangleWarningOutlined";
-import { useTheme } from "@emotion/react";
-import { PriceImpactStatus } from "@hooks/swap/use-swap-handler";
-import { useTranslation } from "react-i18next";
 
 interface ContentProps {
   swapTokenInfo: SwapTokenInfo;
