+++ packages/web/src/views/swap/containers/swap-container/SwapContainer.tsx
@@ -2,11 +2,12 @@ import { useAtomValue } from "jotai";
 import React, { useEffect, useState } from "react";
 
 import { GNOT_TOKEN_DEFAULT } from "@common/values/token-constant";
-import SwapCard from "@components/swap/swap-card/SwapCard";
 import useRouter from "@hooks/common/use-custom-router";
 import { useSwapHandler } from "@hooks/swap/use-swap-handler";
 import { useTokenData } from "@hooks/token/use-token-data";
 import { ThemeState } from "@states/index";
+
+import SwapCard from "../../components/swap-card/SwapCard";
 
 const SwapContainer: React.FC = () => {
   const themeKey = useAtomValue(ThemeState.themeKey);
